+++ apps/router-demo/router-remote2-2002/src/export-App.tsx
@@ -1,6 +1,7 @@
 import App from './App';
-import { createBridgeComponent } from '@module-federation/bridge-react';
+import { createBridgeComponent } from '@module-federation/bridge-react/v18';
 
+// @ts-ignore
 export const provider = createBridgeComponent({
   rootComponent: App,
   defaultRootOptions: {