+++ apps/router-demo/router-remote4-2004/src/export-App.tsx
@@ -1,5 +1,5 @@
 import App from './App';
-import { createBridgeComponent } from '@module-federation/bridge-react';
+import { createBridgeComponent } from '@module-federation/bridge-react/v18';
 
 const provider = createBridgeComponent({
   rootComponent: App,