+++ packages/nextjs-mf/src/types/btoa.d.ts
@@ -1,0 +1,4 @@
+declare module 'btoa' {
+  function btoa(str: string): string;
+  export = btoa;
+}