+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @module-federation/runtime
 
+## 0.11.4
+
+### Patch Changes
+
+- Updated dependencies [64a2bc1]
+- Updated dependencies [c14842f]
+  - @module-federation/sdk@0.11.4
+  - @module-federation/runtime-core@0.11.4
+  - @module-federation/error-codes@0.11.4
+
 ## 0.11.3
 
 ### Patch Changes